+++ spec/models/planned_disbursement_spec.rb
@@ -27,14 +27,4 @@ RSpec.describe PlannedDisbursement, type: :model do
       end
     end
   end
-
-  describe "#unknown_receiving_organisation_type?" do
-    it "returns true when receiving organisation type is 0" do
-      planned_disbursement = create(:planned_disbursement, receiving_organisation_type: "0")
-      expect(planned_disbursement.unknown_receiving_organisation_type?).to be true
-
-      planned_disbursement.update(receiving_organisation_type: "10")
-      expect(planned_disbursement.unknown_receiving_organisation_type?).to be false
-    end
-  end
 end